feat: configured Spring security
* feat: created migrations to work with Spring security

* feat: added dependencies and configured Spring Security

* feat: added AuthAdapter to get authorized user

* refactor: optimized resume dto's

* refactor: improved create, edit and delete logic with using AuthAdapter

* refactor: optimized vacancy dto's

* refactor: improved create, edit and delete logic with using AuthAdapter

* chore: moved AuthAdapter to utils

* feat: added new role admin and data

* refactor: improved logic to get resume by Id

* feat: added sql validation exception handler

* feat: added classes to check role's existence

* refactor: changed everything related to user due to changes in db

* refactor: improved edit and delete logic with using AuthAdapter

* feat: handled Multipart and IllegalState errors

* feat: added classes for responding to vacancies

* feat: handled new category and response exceptions

* all done

---------

+++ src/main/java/kg/attractor/job_search/service/impl/ResumeServiceImpl.java
@@ -3,10 +3,9 @@ package kg.attractor.job_search.service.impl;
 import kg.attractor.job_search.dao.ResumeDao;
 import kg.attractor.job_search.dto.ContactInfoDto;
 import kg.attractor.job_search.dto.EducationInfoDto;
-import kg.attractor.job_search.dto.resume.CreateResumeDto;
-import kg.attractor.job_search.dto.resume.EditResumeDto;
 import kg.attractor.job_search.dto.resume.ResumeDto;
 import kg.attractor.job_search.dto.WorkExperienceInfoDto;
+import kg.attractor.job_search.dto.resume.ResumeFormDto;
 import kg.attractor.job_search.exception.*;
 import kg.attractor.job_search.exception.ApplicantNotFoundException;
 import kg.attractor.job_search.mapper.ResumeMapper;
@@ -16,6 +15,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,6 +26,7 @@ import java.util.List;
 import java.util.function.BiFunction;
 import java.util.function.Function;
 import java.util.function.ObjLongConsumer;
+import java.util.function.ToLongFunction;
 
 @Slf4j
 @Service
@@ -86,12 +87,11 @@ public class ResumeServiceImpl implements ResumeService {
 
     @Override
     @Transactional(rollbackFor = {Exception.class})
-    public Long createResume(CreateResumeDto resumeDto) {
+    public Long createResume(ResumeFormDto resumeDto) {
         if (resumeDto.getApplicantId() == null) {
             throw new ApplicantNotFoundException();
         }
 
-        userService.getApplicantById(resumeDto.getApplicantId());
         categoryService.getCategoryIdIfPresent(resumeDto.getCategoryId());
 
         Long resumeId = resumeDao.createResume(ResumeMapper.toResume(resumeDto));
@@ -122,21 +122,36 @@ public class ResumeServiceImpl implements ResumeService {
     }
 
     @Override
-    public ResumeDto getResumeById(Long resumeId) {
+    public ResumeDto getResumeById(Long resumeId, Long userId) {
+        ResumeDto resume = getResumeById(resumeId);
+        try {
+            userService.getEmployerById(userId);
+            return resume;
+        } catch (EmployerNotFoundException ignore) {
+            if (isResumeOwnedByApplicant(resumeId, userId)) {
+                return resume;
+            }
+            throw new AccessDeniedException("Это не ваше резюме");
+        }
+    }
+
+    private ResumeDto getResumeById(Long resumeId) {
         Resume resume = resumeDao.getResumeById(resumeId)
                 .orElseThrow(() -> new ResumeNotFoundException("Не существует резюме с таким id!"));
 
         ResumeDto resumeDto = ResumeMapper.toResumeDto(resume);
         enrichResumeWithAdditionalData(resumeDto);
-        log.info("Retrieved resume: {}", resumeDto.getId());
         return resumeDto;
     }
 
     @Override
     @Transactional(rollbackFor = {Exception.class})
-    public Long updateResume(Long resumeId, EditResumeDto resumeDto) {
+    public Long updateResume(Long resumeId, ResumeFormDto resumeDto) {
         getResumeById(resumeId);
         categoryService.getCategoryIdIfPresent(resumeDto.getCategoryId());
+        if (!isResumeOwnedByApplicant(resumeId, resumeDto.getApplicantId())) {
+            throw new AccessDeniedException("У вас нет прав на редактирование этого резюме");
+        }
 
         Resume resume = ResumeMapper.toResume(resumeDto);
         resume.setId(resumeId);
@@ -171,8 +186,11 @@ public class ResumeServiceImpl implements ResumeService {
     }
 
     @Override
-    public HttpStatus deleteResume(Long resumeId) {
+    public HttpStatus deleteResume(Long resumeId, Long userId) {
         getResumeById(resumeId);
+        if (!isResumeOwnedByApplicant(resumeId, userId)) {
+            throw new AccessDeniedException("У вас нет прав на удаление этого резюме!");
+        }
         resumeDao.deleteResume(resumeId);
         log.info("Deleted resume: {}", resumeId);
         return HttpStatus.OK;
@@ -191,6 +209,19 @@ public class ResumeServiceImpl implements ResumeService {
         return resumes;
     }
 
+    @Override
+    public Long changeActiveStatus(Long resumeId, Long userId) {
+        ResumeDto resume = getResumeById(resumeId);
+        if (!isResumeOwnedByApplicant(resumeId, userId)) {
+            throw new AccessDeniedException("У вас нет прав на изменение статуса этого резюме!");
+        }
+
+        Boolean status = resume.getIsActive().equals(Boolean.FALSE) ? Boolean.TRUE : Boolean.FALSE;
+
+        resumeDao.updateResumeActiveStatus(resumeId, status);
+        return resumeId;
+    }
+
     private <T> void processResumeItems(Collection<T> items, Long resumeId,
                                         ObjLongConsumer<T> setResumeIdFunc,
                                         Function<T, ?> createFunction) {
@@ -207,7 +238,7 @@ public class ResumeServiceImpl implements ResumeService {
             List<T> newItems,
             ObjLongConsumer<T> setResumeIdFunc,
             BiFunction<Long, T, Long> updateFunction,
-            Function<T, Long> createFunction
+            ToLongFunction<T> createFunction
     ) {
         if (newItems != null) {
             for (T item : newItems) {
@@ -220,17 +251,17 @@ public class ResumeServiceImpl implements ResumeService {
                     if (itemId != null) {
                         updateFunction.apply(itemId, item);
                     } else {
-                        createFunction.apply(item);
+                        createFunction.applyAsLong(item);
                     }
-                } catch (NoSuchMethodException e) {
+                } catch (NoSuchMethodException | InvocationTargetException | IllegalAccessException e) {
                     log.error("Метод getId() не найден для класса {}", item.getClass().getName());
-                } catch (InvocationTargetException e) {
-                    throw new RuntimeException(e);
-                } catch (IllegalAccessException e) {
-                    throw new RuntimeException(e);
                 }
             }
         }
+    }
+
+    public boolean isResumeOwnedByApplicant(Long resumeId, Long applicantId) {
+        return resumeDao.isResumeOwnedByApplicant(resumeId, applicantId);
     }
 
     private void enrichResumesWithAdditionalData(List<ResumeDto> resumes) {
